K8s action support
This commit enables k8s actions by executing them the
same way as any other ns-action. It needs a mandatory
parameter which is application-name.

This commit adds an additional check so that if no
charm is present in the kdu-configuration section
it is understood that the action will be executed
on the KDU workloads.

The commit 8761 in N2VC depends on this commit

https://osm.etsi.org/gerrit/#/c/osm/N2VC/+/8761/

Change-Id: I6adf9a4f786690af5222a208137d95fc28949b34
Signed-off-by: Dominik Fleischmann <dominik.fleischmann@canonical.com>